This change allows to reproduce a logic error that needs to be fixed!
The previous change added padding to the job telemetry if the job started before the simulation.
Original Code:
cpu_trace = np.concatenate((np.array([<idle_power>] * missing_steps),cpu_trace))
This resulted in GW Power as the values should be between 0 and NUM_PROC.

This was fixed by  setting the values to 0:
cpu_trace = np.concatenate((np.array([0] * missing_steps),cpu_trace))

However:
The inial code showed that the error is not only in the number produced, but in
the offset chosen within the cpu_trace (and gpu_trace).

The newly committed code adds
cpu_trace = np.concatenate((np.array([np.NaN] * missing_steps),cpu_trace))
Which triggers this error.
The np.NaN should be in the code an the logic error needs to be fixed!