Updated dataloader to derive cpu and gpu power more accurate per node.
This is a fix, but only a patch-work fix.
There are several assumptions that do not really hold for this dataset:
Power in raps is always between idle and max.
Power here, can be below as other C-States are possible.

Additionally, this dataset provides per node traces, while RAPS assumes one
trace per job, which applies to all nodes, this should be overhauled.
(In this dataset there is one value per node, not one trace per node, though.)