Initial restructure to allow for additional distributions.
TODO: The distributions are not implemented properly.
This sets the initial steps to be able to add the codes and distributions.
Additional work is needed to actually draw from the correct distributions given the parameters.
Complete the set.
Allow for multimodal distributions (Initials are made)
And set default parameters e.g. in a config file that can be specified in the command line.