Added Experimental Flag to map the number of nodes selected to the nearest power of N
Indicated by this Flag:
parser.add_argument("--jobsize-nearest-power-of", default=1, type=int,required=False,help="Map random samples to the nearest power of N your choice. (Experimental: This changes the shape of the distribution, as density of powers change on the numberline!)")

This is not ideal, as the distribution changes. Improved solutions are a welcome contribution.