Removed requested_nodes in favour of using PolicyType to decide if a job is replayed or rescheduled.
Removed requested_nodes for jobs.
These now only have scheduled_nodes.

requested_nodes was used to decide if a job should be rescheduled or replayed.
As this is not the decision of the job but of the scheduling policy, this is not adjusted.

The updated mechanisms use PolicyType to make the decision and if scheduled_nodes is set,
then use these nodes for the replay. If scheduled nodes are not set, place the job on free nodes.

If a PolicyType other than replay is used, place the job freely.

In either case:
Before a job is executed scheduled_nodes indicates where the job was placed according to the telemetry.
If no such information is given, this is fine.
After a job is executed scheduled_nodes indicates where the job was placed by the scheduler.

When implementing to deciding how to place a job use the policy directly.