Added an experimental pytest for workload generation script.
As Workloads plot pyplots this is tested with a timeout.
If no other error occured it. Success is assumed.

In general:
To test everything e.g. for adastsraMI250
pytest -m "adastraMI250 and not cooling and not withdata" -n auto -x

To test everything e.g. for frontier
pytest -m "frontier and not cooling and not withdata" -n auto -x