Shift telemetry_start back to match sim_config.start
With the current dataloaders, WorkloadData.start_date can end up being
after sim_config.start. Most of the dataloaders set timestep_start and
start_date to whatever the first job in the dataset is. This issue is
most apparent when the dataloader does any filtering based on start/end
so start_date will be whichever job started in the interval.

Eventually I think we should refactor the dataloaders to return absolute
unix timestamps to avoid all these relative vs non-relative timestep
complications. But for now I'm just going to shift the telemetry_start
back to match the sim_config.start if needed. This does mean that
telemetry_start can now be a negative value, but that should work fine.