Fixed get_current_utilization to work with trace_quanta set to None again, if trace is an integer or float.
This was a regression, however should be fixed by introducing a trace type containing the information, what the trace is and not rely on the type only. This also would allow to introduce trace types that are e.g. a trace per node of a job. and traces that are timeseries compared to arrays.